+++ Systems/Engine/EngineContainers.hpp
@@ -235,8 +235,8 @@ public:
   template<typename searchKeyType>
   size_t FindApproxIndex(const searchKeyType& key)
   {
-    int begin = 0;
-    int end = (int)mArray.Size();
+    size_t begin = 0;
+    size_t end = (int)mArray.Size();
   
     while(begin < end)
     {
@@ -272,7 +272,7 @@ public:
   
   uint Size()
   {
-    return mArray.Size();
+    return (uint)mArray.Size();
   }
   
   value_type& Front()
