+++ ZeroLibraries/Support/ChunkReader.hpp
@@ -112,14 +112,14 @@ public:
   }
 
   template<typename type>
-  void ReadArray(type* data, uint count)
+  void ReadArray(type* data, size_t count)
   {
     Status status;
     file.Read(status, (byte*)data, sizeof(type) * count);
   }
 
   template<typename type>
-  void ReadArraySize(type* data, uint size)
+  void ReadArraySize(type* data, size_t size)
   {
     Status status;
     file.Read(status, (byte*)data, size);
